+++ part_1/ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: mvidal-a <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/04 18:42:21 by mvidal-a          #+#    #+#             */
-/*   Updated: 2019/12/23 12:05:39 by mvidal-a         ###   ########.fr       */
+/*   Updated: 2020/01/29 16:25:04 by mvidal-a         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 int		ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	unsigned long	i;
-	int				ret;
+	size_t	i;
+	int		ret;
 
 	i = 0;
 	ret = 0;
