+++ part_1/ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: mvidal-a <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/05 18:33:06 by mvidal-a          #+#    #+#             */
-/*   Updated: 2019/11/30 12:56:46 by mvidal-a         ###   ########.fr       */
+/*   Updated: 2020/01/29 16:26:10 by mvidal-a         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,9 +15,9 @@
 
 char	*ft_strdup(const char *s1)
 {
-	char			*dup;
-	size_t			len;
-	unsigned long	i;
+	char	*dup;
+	size_t	len;
+	size_t	i;
 
 	len = ft_strlen(s1);
 	dup = (char *)malloc(sizeof(char) * (len + 1));
